+++ changelog.md
@@ -1,7 +1,21 @@
+# 1.0.2
+- Minor mod Settings bugfixes
+
+# 1.0.1
+- Added 'Active particles' counter
+- Added 'Total triggers' counter
+- Added mod Settings:
+  - Active particles Overlay opacity
+  - Active particles Overlay position
+  - Active particles Overlay size
+  - Total triggers Overlay opacity
+  - Total triggers Overlay position
+  - Total triggers Overlay size
+
+
 # 1.0.0
 - ObjInfo official release!
 - Added mod Settings:
-    - Overlay opacity
-    - Overlay position 
-    - Settings editor
+  - Overlay opacity
+  - Overlay position
 - Having fun seeing level`s optimisation included!